Fix shared mutability in TestTransport (#138)
* Fix shared requests and responses mutability in TestTransport

The problem surfaced when adding multiple responses to TestTransport
and then making consecutive Web3 calls against the test transport. In
this case, the first response would end up being returned to all requests,
not just the first.

The problem was caused by TestTransport storing its mock responses in
a RefCell<VecDeque<json_rpc::Value>>. Thus, whenever the TestTransport
was cloned, a clone of the RefCell was created. This in turn meant that
every TestTransport clone had its own copy of the VecDeque and
mutability wasn't actually shared between these clones as desired.

The solution is simple: add an Rc around the RefCell for both requests
and responses. The Rc will ensure that the same RefCell is shared
between all clones of TestTransport.

* Rename response to responses in TestTransport