Fixed the "replay_block_transactions" method of api::Traces and added tests
The method was calling "trace_replayBlockTransaction" instead of "trace_replayBlockTransactions" (missing "s" at the end) and was returning a single BlockTrace instead of a vector of them (https://wiki.parity.io/JSONRPC-trace-module.html#trace_replayblocktransactions).

Tests were also added for both the method (using the block 46147) and deserialization of vectors of BlockTrace.

Signed-off-by: Matthieu Le brazidec <matthieu@lebrazidec.email>