Allow tokenization of tuples
With solidity abi v2 it has become possible to pass structs directly as
arguments to functions. A struct containing types T0 and T1 is encoded
as a tuple (T0, T1).

In the current state of rust-web3 it is not possible to pass tuples
because of the way the tokenization related traits are set up. We have
Tokenizable which is implemented for basic types like i32 and for arrays
of basic types but not for tuples of Tokenizable. Tokenize is
implemented on tuples of Tokenizable but not tuples of Tokenize so
nesting of tuples is not possible.

This PR allows tokenization of tuple types by implementing Tokenizable
on tuples of Tokenizable. In order for this to not introduce trait
ambiguity (which the compiler also notices) we have to remove the
implementations of Detokenize and Tokenize on a single T: Tokenizable.
This would be ambiguous because it would not be clear whether a (t0, t1)
refers to two function arguments or a single tuple function argument.

This last part means this is a (small) breaking change as function
arguments and return types (for example in Contract::query) now always
have to be tuples (or Vec<Token>) and cannot be single types anymore.

It might be possible to avoid the breakage by doing more trait magic to
still allow non tupled types in those places. They would implicitly be
treated as a tuple of a single element. I feel in the long term this is
more confusing as users would have to remember the special case instead
of always tupling. For example, a function taking a solidity bool would
be callable with a rust bool or (bool,) but a function taking a (bool,)
would only be callable with ((bool,),).

Note that the current release of ethabi does no handle abi v2 but there
are PRs and forks that do which ethabi will hopefully integrate.